% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_camera_coordinates.R
\name{average_camera_coordinates}
\alias{average_camera_coordinates}
\title{Weight camera locations by effort}
\usage{
average_camera_coordinates(locations, effort)
}
\arguments{
\item{locations}{dataframe of camera locations}

\item{effort_by_day}{}
}
\description{
Use the number of day sof effort generated by the effort_by_day function to
average coordinates of nearby cameras with merged locations.
}
