% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_proportion_classified.R
\name{calculate_prop_classified}
\alias{calculate_prop_classified}
\title{Title}
\usage{
calculate_prop_classified(seasons, effort_day, effort_occ)
}
\arguments{
\item{max_year}{}
}
\description{
Title
}
