% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_season_dates.R
\name{create_season_dates}
\alias{create_season_dates}
\title{Create Seasonal Data Frames to Query Detections and Effort}
\usage{
create_season_dates(min_date, max_date, years, discontinuous = FALSE)
}
\arguments{
\item{min_date}{character, start date of query, in the format "-mm-dd"}

\item{max_date}{character, end date of query, in the format "-mm-dd"}

\item{years}{Vector of the start and end years desired. \code{seq()} used within
function to create a sequence of years from start to end years. Change to a
vector of specific years desired if you want a discontinuous sequence of years
AND use argument discontinuous = TRUE.}

\item{discontinuous}{logical indicating whether you want a discontinuous
range of years from which to pull data. Defaults to FALSE. If TRUE, change years
argument to specify the years you want (e.g. c(2019, 2021, 2023)) as opposed
to the start and end years}
}
\value{
Data frame containing years and associated start and end dates
}
\description{
Query effort and detection data from a set of season dates in each year of interest.
}
\examples{
\dontrun{
#continuous year range
create_season_dates(min_date = '-12-01',
                    max_date = '-01-31',
                    years = c('2018','2024'))

#discontinuous year range
dates2 <- create_season_dates(min_date = '-12-01',
                             max_date = '-01-31',
                             years = c('2018','2022','2024'),
                             discontinuous = TRUE)


}
}
