% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_overlap.R
\name{detect_overlap}
\alias{detect_overlap}
\title{Detect Effort Overlap in Batches and Cameras}
\usage{
detect_overlap(df)
}
\arguments{
\item{df}{Effort data frame}
}
\value{
Data frame showing instances of overlap.
}
\description{
Find instances of overlapping effort by camera location and year.
}
\examples{
\dontrun{
effort_df \%>\% detect_overlap()
}
}
