% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_wiscland2.R
\name{download_wiscland2}
\alias{download_wiscland2}
\title{Download and unzip Wiscland2 landcover data from DNR GIS website}
\usage{
download_wiscland2()
}
\value{

}
\description{
Download and unzip Wiscland2 landcover data from DNR GIS website
}
