% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort_query.R
\name{effort_query}
\alias{effort_query}
\title{Pulls in effort data from database}
\usage{
effort_query(conn, prec, grid, daterange = NULL, remove0Timelapse = TRUE)
}
\arguments{
\item{conn}{connection to database returned by \code{sswids::connect_to_sswidb()}}

\item{prec}{precision of classifications. Should be set to 0 or 0.95. 0.95
indicates final data and returns FINAL data while 0 indicates data
that is classified but includes classifications that don't meet
accuracy standards.}

\item{grid}{specified as character vector. What camera grids to pull effort
info for (e.g. "SSWI").}

\item{daterange}{preferably a data frame of start and end dates, with columns
"start_date" and "end_date". Can accommodate same ranges
across years. The date range to pull the effort for. If not
provided defaults to 2019-01-01 to  the current date - 1 year.}

\item{remove0Timelapse}{Logical, should rows with 0 timelapse photos be removed
from the table. These rows indicate camera site/days
where no timelapse photo was taken despite programming.}
}
\value{
a data frame with daily camera effort information as well as location and grid information.
}
\description{
Querys database for information about effort at camera sites,
including classification effort. Combines data from DS_LOCATION_EFFORT,
SSWI_GRID_REF, and SSWI_CAMERA_LOCATION based on specified arguments.
}
