% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/establish_sswidb_credentials.R
\name{establish_sswidb_credentials}
\alias{establish_sswidb_credentials}
\title{Title}
\usage{
establish_sswidb_credentials(db_version)
}
\arguments{
\item{db_version}{}
}
\value{

}
\description{
Title
}
