% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_detections_effort.R
\name{join_detections_effort}
\alias{join_detections_effort}
\title{Title}
\usage{
join_detections_effort(effort_by_occ, detections, discard_na_rows = TRUE)
}
\arguments{
\item{discard_na_rows}{}
}
\value{

}
\description{
Title
}
