% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_classification_method.R
\name{plot_classification_method}
\alias{plot_classification_method}
\title{Title}
\usage{
plot_classification_method(detections)
}
\arguments{
\item{detections}{}
}
\value{

}
\description{
Title
}
