% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_detections.R
\name{plot_detections}
\alias{plot_detections}
\title{Plot detections spatially (by year or across all years)}
\usage{
plot_detections(detections, locations, class_key, years)
}
\arguments{
\item{years}{}
}
\description{
Plot detections spatially (by year or across all years)
}
