% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_detections.R
\name{query_detections}
\alias{query_detections}
\title{Pull in detection from Snapshot database. A wrapper around \code{sswidb::sswidb_detections()}.}
\usage{
query_detections(conn, species, grid, daterange = NULL, prec, county = NULL)
}
\arguments{
\item{conn}{a connection to the sswi database}

\item{species}{character vector of species of interest.See \code{sswidb::sswidb_species}, species=NULL returns all species}

\item{grid}{specified as character vector. What camera grids to pull effort
info for (e.g. "SSWI").}

\item{daterange}{preferably a data frame of start and end dates, with columns
"start_date" and "end_date". Can accommodate same ranges
across years. The date range to pull the effort for. If not
provided defaults to 2019-01-01 to  the current date - 1 year.}

\item{prec}{precision of classifications. Should be set to 0 or 0.95. 0.95
indicates final data and returns FINAL data while 0 indicates data
that is classified but includes classifications that don't meet
accuracy standards.}

\item{county}{a county name filter, uses all counties when NULL, passed on to \code{sswidb::sswidb_detections()}}
}
\description{
Pull in detection from Snapshot database. A wrapper around \code{sswidb::sswidb_detections()}.
}
