% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_bad_batches.R
\name{rm_bad_batches}
\alias{rm_bad_batches}
\title{Removes impossible dates outside the start and end active dates of a batch}
\usage{
rm_bad_batches(data = list_nolocs)
}
\arguments{
\item{listobj}{}
}
\description{
At least some of these are errors because of the location start date not
overlapping with the batch start and end dates. The batch START_ACTIVE_DATE
and END_ACTIVE_DATE and the location START_DATE and END_DATE are both assessed
when determining start_date and end_date for effort from the sswidb_effort
function.Find all batches that have detections outside of batch start and end
active dates and remove those batches and all associated detections and locations.
}
