% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_bad_locations.R
\name{rm_bad_locations}
\alias{rm_bad_locations}
\title{Remove known bad locations}
\usage{
rm_bad_locations(locationeffort = locationeffort, coordinate_precision)
}
\arguments{
\item{locationeffort}{data frame of camera location and effort data from \code{effort_query()}}
}
\value{
a data frame with known bad locations removed.
}
\description{
Removes bad locations, specifically one camera that is actually in the UP of
Michigan and another with an erroneous location in the middle of a lake.
}
