% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_noloc_data.R
\name{rm_noloc_data}
\alias{rm_noloc_data}
\title{Remove data without location information}
\usage{
rm_noloc_data(
  locs_df = locs_df_raw,
  effort_df = effort_df_raw,
  detections_df = detections_df_raw,
  spatial_data = "counties",
  zone = county_name
)
}
\arguments{
\item{locs_df}{raw locations dataframe from Snapshot database query}

\item{effort_df}{raw effort dataframe from Snapshot database query}

\item{detections_df}{raw detections dataframe from Snapshot database query}

\item{spatial_data}{character string returned by list_spatial_layers()}

\item{zone}{column name (unquoted!), specify column of spatial_data that contains zone or management unit id by which to identify locations outside the study area}
}
\value{
a list containing filtered locations dataframe with UP and no location/effort data , filtered effort dataframe with UP and no location/effort data, filtered detection dataframe with UP and no location/effort data
}
\description{
Removes detections without Wisconsin location data. There are camera_location_seq_no's
in effort and detections dataframes that have no location data.
This is likely because of detection date times outside of the camera_location_seq_no
start and end dates.Also, there is 1 location that falls in the UP of Michigan.
Remove this location as well. This is an old
location from 2020 and earlier, so it cannot be corrected.
}
