% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_project.R
\name{setup_project}
\alias{setup_project}
\title{Create project directories to store SSWI data and outputs}
\usage{
setup_project()
}
\value{

}
\description{
Create project directories to store SSWI data and outputs
}
