% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_subset.R
\name{spatial_subset}
\alias{spatial_subset}
\title{Spatially Subset Camera Site Locations}
\usage{
spatial_subset(df, polygon_layer)
}
\arguments{
\item{df}{Data frame of camera locations}

\item{polygon_layer}{sf polygon object}
}
\value{
data frame
}
\description{
Spatially subset camera site locations using an sf polygon object.
}
\examples{
\dontrun{
spatial_subset(locs_df, _sf)
}
}
