% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_sswidb_tables.R
\name{sswidb_tables}
\alias{sswidb_tables}
\title{Print Snapshot Wisconsin Database Tables}
\usage{
sswidb_tables()
}
\value{
list of Snapshot Wisconsin tables
}
\description{
Print a list of Snapshot Wisconsin tables from the larger WI DNR Oracle database
}
\examples{
\dontrun{
# first connect to Snapshot Wisconsin database
connect_to_sswidb(db_version = 'PROD')
# print Snapshot Wisconsin tables
sswidb_tables()
}
}
