% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_by_occasion.R
\name{summarize_by_occasion}
\alias{summarize_by_occasion}
\title{Summarize by Occasion}
\usage{
summarize_by_occasion(dataframe, yearadd)
}
\arguments{
\item{yearadd}{}
}
\description{
Summarize by Occasion
}
