% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFunctions.R
\name{tile}
\alias{tile}
\title{Bin a vector of dates into occasions of defined length}
\usage{
tile(x, len)
}
\arguments{
\item{x}{The vector of dates to be binned}

\item{len}{desired occasion length}
}
\value{
vector of numeric occasion assignments. If total interval is not equally divisible by len,
then remainder days are either assigned into the final complete occasion (if remainder/len <0 .5),
in which case the final occasion is longer than specified len or else
assigned into an additional occasion shorter then len, if remainder/len >= 0.5.
}
\description{
Bin a vector of dates into occasions of defined length
}
\examples{
\dontrun{
a <- seq(as.Date("2020-10-01"),as.Date("2020-12-31"),1)
tile(a,7)
}
}
