% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_trigger.R
\name{view_trigger}
\alias{view_trigger}
\title{Look at a set of triggers in .gif format}
\usage{
view_trigger(trigger_seq_no)
}
\arguments{
\item{trigger_seq_no}{trigger sequence number}
}
\description{
Look at a set of triggers in .gif format
}
\examples{
\dontrun{
view_trigger(trigger_seq_no = 21733907)
}
}
