% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_encounter_histories.R
\name{write_encounter_histories}
\alias{write_encounter_histories}
\title{Write Encounter Histories for Abundance Analysis}
\usage{
write_encounter_histories(effort_by_occ)
}
\arguments{
\item{df}{}
}
\value{
csv files of encounter histories by class (e.g., age and sex)
}
\description{
Write Encounter Histories for Abundance Analysis
}
\examples{
\dontrun{
write_encounter_histories(df)
}
}
