% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_camera_coordinates.R
\name{average_camera_coordinates}
\alias{average_camera_coordinates}
\title{Title}
\usage{
average_camera_coordinates(locations, effort_by_day)
}
\arguments{
\item{effort_by_day}{}
}
\description{
Title
}
