% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_camera_coordinates.R
\name{average_camera_coordinates}
\alias{average_camera_coordinates}
\title{Title}
\usage{
average_camera_coordinates(locations, effort)
}
\arguments{
\item{effort}{}
}
\value{

}
\description{
Title
}
