% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sampling_occasions.R
\name{create_sampling_occasions}
\alias{create_sampling_occasions}
\title{Title}
\usage{
create_sampling_occasions(seasons, effort, num_occasions)
}
\arguments{
\item{num_occasions}{}
}
\value{

}
\description{
Title
}
