% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_season_dates.R
\name{create_season_dates}
\alias{create_season_dates}
\title{Create Seasonal Data Frames to Query Detections and Effort}
\usage{
create_season_dates(min_date, max_date, years, discontinuous = FALSE)
}
\arguments{
\item{min_date}{Start date of query}

\item{max_date}{End date of query}

\item{years}{Vector of the start and end years desired. \code{seq()} used within
function to create a sequence of years from start to end years. Change to a
vector of specific years desired if you want a discontinuous sequence of years.}

\item{discontinuous}{logical indicating whether you want a discontinuous
range of years from which to pull data. Defaults to FALSE. If TRUE, change years
argument to specify the years you want (e.g. c(2019, 2021, 2023)) as opposed
to the start and end years}
}
\value{
Data frame containing years and associated start and end dates
}
\description{
Query effort and detection data from a set of season dates in each year of interest.
}
\examples{
\dontrun{
# set min/max years for obtaining data
min_year <- 2017
max_year <- 2022

# season start/end dates (in -MM-DD format)
min_date <- '-04-01'
max_date <- '-04-30'

# what years of data to query?
years <- seq(min_year, max_year, 1)

# create date ranges for each year;
# if spanning two years (i.e., winter) use start date year
# otherwise just use the year of the particular season
seasons_df <-
create_season_dates(
min_date,
max_date,
years
)
}
}
