% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/establish_sswidb_credentials.R
\name{establish_sswidb_credentials}
\alias{establish_sswidb_credentials}
\title{Set Snapshot Wisconsin database credentials just one time for future database connections}
\usage{
establish_sswidb_credentials(db_version)
}
\arguments{
\item{db_version}{Production ("PROD") or testing ("UAT") versions of the database.}
}
\value{
Confidential information saved in the credential store.
}
\description{
Use \code{\link{keyring}} package to store Snapshot Wisconsin database credentials for
future database connections.
}
\examples{
\dontrun{
# store credentials for production version of the Snapshot Wisconsin database
# this will create 3 pop-ups that require entry of 1) dsn, 2) uid, and 3) pwd
establish_sswidb_credentials(db_version = 'PROD')
}
}
\seealso{
\code{\link{connect_to_sswidb}}
}
