% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_nearby_cameras.R
\name{merge_nearby_cameras}
\alias{merge_nearby_cameras}
\title{Merge nearby camera_location_seq_no's and create new camera locations called cam_site_id's}
\usage{
merge_nearby_cameras(locs_df, cam_distance, effort_df)
}
\arguments{
\item{effort_df}{}
}
\value{

}
\description{
Merge nearby camera_location_seq_no's and create new camera locations called cam_site_id's
}
