% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_bad_locations.R
\name{rm_bad_locations}
\alias{rm_bad_locations}
\title{Remove known bad locations}
\usage{
rm_bad_locations(locationeffort = locationeffort, coordinate_precision = 4)
}
\arguments{
\item{locationeffort}{data frame of camera location and effort data from \code{query_effort()}}

\item{coordinate_precision}{numeric, number of decimal places to filter locations by. Default is 4.}
}
\value{
a data frame with known bad locations removed.
}
\description{
Removes bad locations, specifically one camera that is actually in the UP of
Michigan and another with an erroneous location in the middle of a lake, as well as
imprecise locations as specified by the user. Imprecise locations are determined by
the number of digits after the decimal. The default is 4 which corresponds to an
accuracy of ~11m. Number of digits after the decimals scales 10x with each
added decimal place.Locations are considered precise if either the latitude OR
longitude coordinate has 4 decimal places (accounting for truncation of 0 as
last decimal place).
}
\examples{
\dontrun{
Q2.test.typical <- rm_bad_locations(locationeffort = Q.test.typical,
                                    coordinate_precision = 4)
}


}
