% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_plot.R
\name{spatial_plot}
\alias{spatial_plot}
\title{Spatial Plots}
\usage{
spatial_plot(
  conn,
  df,
  mgmtlayer = "counties",
  days_active_threshold,
  ppn_class_threshold,
  n_occasions_annual_threshold,
  spatialgroup = "COUNTY_NAM",
  combine_cols = TRUE,
  layer2 = NULL
)
}
\arguments{
\item{conn}{connection to the Snapshot database from \code{connect_to_sswidb()}}

\item{df}{a sf data frame output by the data pull workflow}

\item{mgmtlayer}{spatial layer from \code{list_spatial_layers()}, defaults to counties}

\item{days_active_threshold}{Numeric, scalar. Minimum number of days a camera needs to have been functioning within an occasion for a cam site id x year x occasion to be included in spatial plot.}

\item{ppn_class_threshold}{Numeric, scalar. Proportion of photos classified within an occasion required for a cam site id x year x occasion to be included in a spatial plot.}

\item{n_occasions_annual_threshold}{Numeric, scalar. Minimum number of occasions required for a cam site x year to be included in a spatial plot.}

\item{spatialgroup}{character, column name in mgmtlayer that denotes either the zone names or county names to summarize camera data by. Defaults to COUNTY_NAM.}

\item{combine_cols}{logical, should species age/sex columns be summed together? Defaults to TRUE}

\item{layer2}{character, specifying a 2nd spatial layer from \code{list_spatial_layers()} to add to plot if mgmtlayer == counties}
}
\value{
a named list of ggplot objects
}
\description{
Produces a list of spatial plots summarizing animal detections by a chosen spatial layer.
Can't currently handle multiple spatial layers within function, would have to run function
for each different spatial layer. It's possible to add to these objects for further modification.
}
