% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spatial.R
\name{spatial_plot}
\alias{spatial_plot}
\title{Spatial Plots}
\usage{
spatial_plot(
  specieslist,
  x,
  mgmtlayer,
  days_active_threshold,
  ppn_class_threshold,
  n_occasions_annual,
  spatialgroup
)
}
\arguments{
\item{specieslist}{a vector of species names for which you want to plot spatial data}

\item{x}{a data frame output by the data pull workflow}

\item{mgmtlayer}{spatial layer from \code{list_spatial_layers()}}

\item{days_active_threshold}{Numeric, scalar. Minimum number of days a camera needs to have been functioning within an occasion for a cam site id x year x occasion to be included in spatial plot.}

\item{ppn_class_threshold}{Numeric, scalar. Proportion of photos classified within an occasion required for a cam site id x year x occasion to be included in a spatial plot.}

\item{n_occasions_annual}{Numeric, scalar. Minimum number of occasions required for a cam site x year to be included in a spatial plot.}

\item{spatialgroup}{character, column name in mgmtlayer that denotes either the zone names or county names to summarize camera data by.}
}
\value{
a named list of ggplot objects
}
\description{
Spatial Plots
}
