% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_subset.R
\name{spatial_subset}
\alias{spatial_subset}
\title{Spatially subset camera site locations using an sf polygon object}
\usage{
spatial_subset(df, polygon_layer)
}
\arguments{
\item{polygon_layer}{}
}
\value{

}
\description{
Spatially subset camera site locations using an sf polygon object
}
