% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_subset.R
\name{spatial_subset}
\alias{spatial_subset}
\title{Spatially Subset Camera Site Locations}
\usage{
spatial_subset(locationeffort, sf_layer, zone_id, filter_statement = NULL)
}
\arguments{
\item{locationeffort}{dataframe containing location and effort information from \code{effort_query()}}

\item{sf_layer}{as character string, sf polygon object by which to subset locationeffort data frame}

\item{zone_id}{as character string, ID column in sf_layer e.g.(bear_mgmt_unit_id)}

\item{filter_statement}{character string of right hand side of filter statment, to be passed along to subset spatial layer (e.g. "== 'A'")}
}
\value{
data frame
}
\description{
Spatially subset camera site locations using an sf polygon object. There may
be times where it doesn’t make sense to include data from the entire Snapshot
Wisconsin camera network, and we should spatially subset camera locations. For
example, we may want to only consider the Northern Forest Zone instead of the
entire state of Wisconsin. We can filter locations, detections, and effort
based on a spatial layer, like deer management unit.
}
\examples{
\dontrun{
spatial_subset(locs_df, _sf)
}
}
\seealso{
\code{list_spatial_layers()}
}
