% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_sswidb_tables.R
\name{sswidb_tables}
\alias{sswidb_tables}
\title{Print a list of SSWI database tables from larger DNR database}
\usage{
sswidb_tables()
}
\value{
list of SSWI tables
}
\description{
Print a list of SSWI database tables from larger DNR database
}
\examples{
sswidb_tables()
}
