% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_trigger.R
\name{view_trigger}
\alias{view_trigger}
\title{Look at a set of triggers in .gif format}
\usage{
view_trigger(trigger_id)
}
\arguments{
\item{trigger_id}{numeric trigger ID}
}
\value{

}
\description{
Look at a set of triggers in .gif format
}
\examples{
\dontrun{
view_trigger(21733907)
}
}
